refactor(gitinfo): extract per-operation handlers from handleModalResult
Extract each pendingOp case from the 323-line handleModalResult switch
into a named private method in gitinfo_modal_handlers.go. The switch
body is now pure routing with single-line delegations.

Introduce modalArgs struct to pass the captured pending state snapshot
to each handler, keeping signatures uniform.

No behavior change — pure refactoring.

Closes #131
